requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
psycopg2-binary ==2.9.6
Flask-Migrate ==4.0.4
